Move logic into common.rs and make everything in lib.rs protected by agave-unstable-api.
