+++ votor/src/consensus_pool/stats.rs
@@ -2,7 +2,7 @@
 // TODO(wen): remove allow(dead_code) when consensus_pool is fully integrated
 
 use {
-    crate::VoteType,
+    crate::common::VoteType,
     solana_metrics::datapoint_info,
     solana_votor_messages::consensus_message::CertificateType,
     std::time::{Duration, Instant},
